refactor: class형태로 `input`과 `racingCarGame` 그리고 `winner`로 쪼개어 관리
